git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK9@1790133 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/test/java/org/apache/sis/util/collection/DerivedSetTest.java
@@ -43,7 +43,7 @@ public final strictfp class DerivedSetTest extends TestCase implements ObjectCon
     /**
      * The value to replace by {@code null}.
      */
-    protected static final int EXCLUDED = 19; // non-private for javadoc purpose.
+    protected static final int EXCLUDED = 19;                   // non-private for javadoc purpose.
 
     /**
      * Tests {@link DerivedSet} without excluded value.
@@ -102,7 +102,7 @@ public final strictfp class DerivedSetTest extends TestCase implements ObjectCon
      */
     @Override
     public Integer apply(final Integer value) {
-        if (value.intValue() == EXCLUDED) {
+        if (value == EXCLUDED) {
             return null;
         }
         return value * 10;
